+++ spec/grape/middleware/exception_spec.rb
@@ -158,13 +158,14 @@ describe Grape::Middleware::Error do
     it 'is possible to specify a custom formatter' do
       @app ||= Rack::Builder.app do
         use Spec::Support::EndpointFaker
-        use Grape::Middleware::Error, rescue_all: true,
-                                      format: :custom,
-                                      error_formatters: {
-                                        custom: lambda do |message, _backtrace, _options, _env|
-                                          { custom_formatter: message }.inspect
-                                        end
-                                      }
+        use Grape::Middleware::Error,
+            rescue_all: true,
+            format: :custom,
+            error_formatters: {
+              custom: lambda do |message, _backtrace, _options, _env, _original_exception|
+                { custom_formatter: message }.inspect
+              end
+            }
         run ExceptionSpec::ExceptionApp
       end
       get '/'
@@ -191,6 +192,47 @@ describe Grape::Middleware::Error do
       get '/'
       expect(last_response.status).to eq(400)
       expect(last_response.body).to eq('failed validation')
+    end
+
+    context 'with rescue_options :backtrace and :exception set to true' do
+      it 'is possible to return the backtrace and the original exception in json format' do
+        @app ||= Rack::Builder.app do
+          use Spec::Support::EndpointFaker
+          use Grape::Middleware::Error,
+              rescue_all: true,
+              format: :json,
+              rescue_options: { backtrace: true, original_exception: true }
+          run ExceptionSpec::ExceptionApp
+        end
+        get '/'
+        expect(last_response.body).to include('error', 'rain!', 'backtrace', 'original_exception', 'RuntimeError')
+      end
+
+      it 'is possible to return the backtrace and the original exception in xml format' do
+        @app ||= Rack::Builder.app do
+          use Spec::Support::EndpointFaker
+          use Grape::Middleware::Error,
+              rescue_all: true,
+              format: :xml,
+              rescue_options: { backtrace: true, original_exception: true }
+          run ExceptionSpec::ExceptionApp
+        end
+        get '/'
+        expect(last_response.body).to include('error', 'rain!', 'backtrace', 'original-exception', 'RuntimeError')
+      end
+
+      it 'is possible to return the backtrace and the original exception in txt format' do
+        @app ||= Rack::Builder.app do
+          use Spec::Support::EndpointFaker
+          use Grape::Middleware::Error,
+              rescue_all: true,
+              format: :txt,
+              rescue_options: { backtrace: true, original_exception: true }
+          run ExceptionSpec::ExceptionApp
+        end
+        get '/'
+        expect(last_response.body).to include('error', 'rain!', 'backtrace', 'original exception', 'RuntimeError')
+      end
     end
   end
 end